[HUDI-1295] Metadata Index - Bloom filter and Column stats index to speed up index lookups
 - HoodieBloomMetaIndexBatchCheckFunction cleanup to remove unused classes

 - Base file checking before reading the file footer for bloom or column stats
