+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/config/HoodieWriteConfig.java
@@ -1440,8 +1440,7 @@ public class HoodieWriteConfig extends HoodieConfig {
   }
 
   public boolean isMetaIndexColumnStatsForAllColumns() {
-    return isMetadataTableEnabled() && isMetaIndexColumnStatsForAllColumns()
-        && getMetadataConfig().isMetaIndexColumnStatsForAllColumns();
+    return isMetadataTableEnabled() && getMetadataConfig().isMetaIndexColumnStatsForAllColumns();
   }
 
   public boolean isMetaIndexBloomFilterBatchLoadEnabled() {
